optimization: Enable short-circuiting for zoom check in should_tmux_control
This change cuts out the `tmux` CLI invocation when
`disable_nav_when_zoomed = false`.

This reordering was necessary to enable this workaround because the
`and` operator is short-circuiting, and thus it will only evaluate the
RHS if the LHS is `true`. Before, the slow function call was the LHS,
and so it was being evaluated unconditionally, even when the
significantly faster boolean check would have failed.