optimization: Circumvent shell for `tmux` CLI invocations
Shell startup time is significant. For some reason, as @isaksamsten
mentioned on #16, switching is significantly slower in Fish shell than
in other shells. I did some experimentation and found that
`vim.fn.system(<string>)` is slow when the `shell` option is set to
Fish. This change removes dependence on the shell entirely when
switching tmux panes, since `system` doesn't use a shell when a table is
provided in place of a string.